requirements.txt
requests
pymongo
Blueprint
gunicorn===19.9.0
pycryptodome
python-dateutil
iso8601
Django
psycopg2-binary
django-rest-swagger
djangorestframework
pylint
pre-commit
autopep8
pylint-django
influxdb
numpy
pandas
sklearn
gevent
urllib3
pyyaml
kafka-python
statsmodels
